:truck: [#17] move the api token files to the api app
