+++ src/woo_publications/api/permissions.py
@@ -2,10 +2,13 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
-from rest_framework.permissions import BasePermission
+from rest_framework.permissions import SAFE_METHODS, BasePermission
 from rest_framework.request import Request
 
 from woo_publications.api.drf_spectacular.headers import ALL_AUDIT_PARAMETERS
+
+from .constants import PermissionOptions
+from .models import TokenAuth
 
 if TYPE_CHECKING:
     from rest_framework.views import APIView
@@ -17,3 +20,25 @@ class AuditHeaderPermission(BasePermission):
             parameter.name for parameter in ALL_AUDIT_PARAMETERS if parameter.required
         ]
         return all(header in request.headers for header in required_headers)
+
+
+class TokenAuthPermission(BasePermission):
+    def has_permission(self, request, view) -> bool:
+        token: TokenAuth = request.auth
+
+        if not token:
+            return False
+
+        if (
+            request.method in SAFE_METHODS
+            and PermissionOptions.read in token.permissies
+        ):
+            return True
+
+        if (
+            request.method not in SAFE_METHODS
+            and PermissionOptions.write in token.permissies
+        ):
+            return True
+
+        return False
